+++ zfit/minimize.py
@@ -11,12 +11,10 @@ from .minimizers.minimizer_nlopt import (NLoptBaseMinimizerV1, NLoptBOBYQAV1,
                                          NLoptShiftVarV1, NLoptSLSQPV1,
                                          NLoptStoGOV1, NLoptSubplexV1,
                                          NLoptTruncNewtonV1)
-from .minimizers.minimizers_scipy import (ScipyBaseMinimizerV1, ScipyCOBYLAV1,
-                                          ScipyLBFGSBV1, ScipyNelderMeadV1,
-                                          ScipyNewtonCGV1, ScipyPowellV1,
-                                          ScipySLSQPV1, ScipyTruncNCV1,
-                                          ScipyTrustConstrV1,
-                                          ScipyTrustKrylovV1, ScipyTrustNCGV1)
+from .minimizers.minimizers_scipy import (ScipyBaseMinimizerV1, ScipyLBFGSBV1,
+                                          ScipyNelderMeadV1, ScipyNewtonCGV1,
+                                          ScipyPowellV1, ScipySLSQPV1,
+                                          ScipyTruncNCV1, ScipyTrustConstrV1)
 from .minimizers.optimizers_tf import Adam, WrapOptimizer
 from .minimizers.strategy import (DefaultToyStrategy, PushbackStrategy,
                                   ZfitStrategy)
@@ -65,7 +63,7 @@ class BFGS:
 __all__ = ["WrapOptimizer",
            "Adam", "Minuit",
            "ScipyBaseMinimizerV1",
-           "ScipyLBFGSBV1", "ScipyTrustKrylovV1", 'ScipyTrustConstrV1', "ScipyTrustNCGV1",
+           "ScipyLBFGSBV1", 'ScipyTrustConstrV1',
            "ScipyPowellV1", "ScipySLSQPV1", "ScipyNewtonCGV1", "ScipyTruncNCV1", "ScipyNelderMeadV1",
            "NLoptBaseMinimizerV1",
            "NLoptLBFGSV1", "NLoptTruncNewtonV1", "NLoptSLSQPV1", "NLoptMMAV1", "NLoptCCSAQV1", 'NLoptShiftVarV1',